# cache: optimize ApplyModifications() set iteration