Lift the new value access out of the hot loop and save
a bunch of time iterating over large sets.

```
Before
---
BenchmarkTableCacheApplyModificationsSet-8   	     268	   4664892 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     266	   4501507 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     229	   4500050 ns/op

BenchmarkTableCacheApplyModificationsSet-8   	     268	   4466431 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     271	   4645624 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     276	   4416883 ns/op

After
---
BenchmarkTableCacheApplyModificationsSet-8   	     520	   2280201 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     480	   2581277 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     469	   2371769 ns/op

BenchmarkTableCacheApplyModificationsSet-8   	     494	   2492495 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     522	   2666051 ns/op
BenchmarkTableCacheApplyModificationsSet-8   	     489	   2602648 ns/op
```