@dcbw 2022-06-01
> LGTM, wow didn't know this could make so much difference, out of curiosity a dumb question: how did you trigger these benchmarks `BenchmarkTableCacheApplyModificationsSet-8`? (it comes from the test, but did you get it from the CI here?)

`make bench` actually, like the golang testing stuff, it'll run any function in a _test.go file that starts with `Benchmark`